1) Переименованы некоторые переменные 2) Добавлено перекрытие (на длину сигнатуры) при зачитке файла по частям

+++ Antivirus/Sequence.h
@@ -24,6 +24,6 @@ struct SequenceData
 //   }
 
   char* Sequence;
-  DWORD SeqLength;
+  LONG SeqLength;
 };
